Add PutNX method on bucket
We added a PutNX (put-if-not-exists) method on the bucket type. This let's you
put a key/val if key doesn't already exist.  The value is not updated if key
already exists.

Note that this is analogous to redis's [SETNX](http://redis.io/commands/setnx).

This was added as a response to the feature request in [issue
2](https://github.com/joyrexus/buckets/issues/2).